# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger API для бибилотеки песен",
        "title": "Онлайн библиотека песен",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/add": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Songs"
                ],
                "summary": "Добавление новой песни",
                "parameters": [
                    {
                        "description": "Данные для добавления песни",
                        "name": "song",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddSongReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные песни",
                        "schema": {
                            "$ref": "#/definitions/model.Song"
                        }
                    },
                    "400": {
                        "description": "Неверное тело запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка добавления песни на сервере",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/songs": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Songs"
                ],
                "summary": "Получение данных библиотеки с фильтрацией по всем полям и пагинацией",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по названию песни",
                        "name": "song",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по названию исполнителя",
                        "name": "group",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Фильтр по дате выпуска",
                        "name": "release_date",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Количество песен на странице",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Номер страницы",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Song"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный фильтр или параметры пагинации",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка получения данных",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/{songID}": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Songs"
                ],
                "summary": "Изменение данных песни",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID песни",
                        "name": "songID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для изменения",
                        "name": "song",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateSongReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные песни изменены",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверное тело запроса или ID песни",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка изменения песни на сервере",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Songs"
                ],
                "summary": "Удаление песни",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID песни",
                        "name": "songID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Песня удалена",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный ID песни",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка удаления песни",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/{songID}/lyrics": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Songs"
                ],
                "summary": "Получение текста песни с пагинацией по куплетам",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID песни",
                        "name": "songID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 3,
                        "description": "количество куплетов на странице",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "номер страницы",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Verse"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный ID песни или некорректные параметры пагинации",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка получения текста песни",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.AddSongReq": {
            "type": "object",
            "required": [
                "group",
                "song"
            ],
            "properties": {
                "group": {
                    "type": "string"
                },
                "song": {
                    "type": "string"
                }
            }
        },
        "dto.UpdateSongReq": {
            "type": "object",
            "properties": {
                "group": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "releaseDate": {
                    "type": "string"
                },
                "song": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "model.Song": {
            "type": "object",
            "required": [
                "group",
                "song"
            ],
            "properties": {
                "group": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "link": {
                    "type": "string"
                },
                "release_date": {
                    "type": "string"
                },
                "song": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "model.Verse": {
            "type": "object",
            "properties": {
                "lyrics": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                }
            }
        }
    }
}